Practice Location Change + Procedure Labels
Changed to practice choiceset locations from .4 (positive and negative directions) to .35 (positive and negative directions. I did this for shapes and text in choice and outcome.

I also added labels to the script to indicate where I am in the script so I am no longer confused about what I am looking at and what it refers to.