Added routineForceEnded and a Trial Function
The issue for routineForceEnded was probably the version of psychopy it was originally run on (2022.2.1). It did not include that code or work regardless when clicking the force routine to check. I also added a trial function for all the different trials (practice, static, dynamic). This way I can change the trials at the beginning without having to go into the script far down.