Pupil Models interacting WMC and Trial
As a simple model, there seems to be an interaction effect of trial with continuous (but not categorical) wmc in both window 2 and 4. However, when including (continuous) current difficulty, patterns slightly change. In window 2, only continuous wmc kept its two-way effect with trial, plus it significantly interacted with current difficulty. It also had a marginal three-way interaction effect