scaling alpha and gamma
original formula leaves alpha to move in not practical and meaningful ways. now alpha is bound to wmc and is more interpretable when it moves laterally.

optim had difficulty with gamma when its values became larger, but by scaling it to its exponential, we can give it something it can handle much more easily without becoming biased. We can then take the output exponential to extract what it meaningfully represents (i.e., the larger value)